update module flower.sample
