+++ flower.sample/src/main/java/com/ly/train/flower/sample/springboot/controller/IndexController.java
@@ -15,27 +15,28 @@
  */
 package com.ly.train.flower.sample.springboot.controller;
 
-import static org.springframework.web.bind.annotation.RequestMethod.POST;
 import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+
 import com.ly.train.flower.common.annotation.Flower;
+import com.ly.train.flower.sample.springboot.service.ServiceA;
+import com.ly.train.flower.sample.springboot.service.ServiceB;
 import com.ly.train.flower.web.spring.FlowerController;
 
 @RestController
-@Flower(serviceName = "ServiceA", value = "async")
+@Flower(value = "async-index", flowNumber = 128)
 public class IndexController extends FlowerController {
 
+	@RequestMapping(value = "/", method = RequestMethod.POST)
+	public void index(Object param, HttpServletRequest req) throws Throwable {
+		doProcess(param, req);
+	}
 
-  @RequestMapping(value = "/", method = POST)
-  public void index(Object param, HttpServletRequest req) throws Throwable {
-    doProcess(param, req);
-  }
-
-
-  @Override
-  public void buildFlower() {
-    getServiceFlow().buildFlow("ServiceA", "ServiceB");
-  }
-
+	@Override
+	public void buildFlower() {
+		getServiceFlow().buildFlow(ServiceA.class, ServiceB.class);
+	}
 }
