+++ flower.sample/src/main/java/com/ly/train/flower/sample/web/FlowServlet.java
@@ -17,76 +17,37 @@ package com.ly.train.flower.sample.web;
 
 import java.io.IOException;
 import java.io.PrintWriter;
+
 import javax.servlet.AsyncContext;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
 import com.ly.train.flower.core.akka.router.FlowRouter;
-import akka.actor.ActorRef;
-import akka.actor.ActorSystem;
-import akka.actor.Props;
+import com.ly.train.flower.core.service.impl.NothingService;
 
 public class FlowServlet extends FlowerHttpServlet {
-  /**
-   * 
-   */
-  private static final long serialVersionUID = 1L;
-  FlowRouter sr;
-  ActorRef actor;
+	private static final long serialVersionUID = 1L;
+	private FlowRouter sr;
+	private final String flowName = "flower-service-flow";
 
-  @Override
-  public void init() {
-    buildServiceEnv();
-    sr = flowerFactory.getServiceFacade().buildFlowRouter("flow", 400);
-    actor = ActorSystem.create("sample").actorOf(Props.create(RouterActor.class));
-  }
+	@Override
+	public void init() {
+		flowerFactory.getServiceFactory().registerService("flowService", FlowService.class);
+		flowerFactory.getServiceFactory().registerService("endService", NothingService.class);
+		getServiceFlow(flowName).buildFlow("flowService", "endService").build();
+		this.sr = flowerFactory.getServiceFacade().buildFlowRouter(flowName, 400);
+	}
 
-  @Override
-  protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-    resp.setContentType("text/html;charset=UTF-8");
-    PrintWriter out = resp.getWriter();
-    out.println("begin：" + System.currentTimeMillis());
-    out.flush();
+	@Override
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		resp.setContentType("text/html;charset=UTF-8");
+		PrintWriter out = resp.getWriter();
+		out.println("begin：" + System.currentTimeMillis());
+		out.flush();
 
-    AsyncContext ctx = req.startAsync();
+		AsyncContext ctx = req.startAsync();
+		sr.asyncCallService(" Hello, Flow World! ", ctx);
+	}
 
-    // flower
-    asyncExe(ctx);
-
-    // thread
-    // new Thread(new Executor(ctx)).start();
-
-    // actor
-    // ServiceContext serviceContext = new ServiceContext();
-    // Web web = new Web(ctx);
-    // serviceContext.setWeb(web);
-    // String uuid = UUID.randomUUID().toString();
-    // FlowContext.putServiceContext(uuid, serviceContext);
-    // actor.tell(uuid, null);
-
-    // out.println("结束Servlet的时间：" + new Date() + ".");
-    // out.flush();
-  }
-
-  private void asyncExe(AsyncContext ctx) {
-    try {
-      sr.asyncCallService(" Hello, Flow World! ", ctx);
-      // ctx.getResponse().getWriter().println("结束Servlet的时间：" + new Date() +
-      // ".");
-      // ctx.complete();
-      // ServiceFacade.asyncCallService("flow", "flowService", " Hello World! ",
-      // ctx);
-    } catch (Exception e) {
-      e.printStackTrace();
-    }
-  }
-
-  private void buildServiceEnv() {
-    flowerFactory.getServiceFactory().registerService("flowService", "com.ly.train.flower.sample.web.FlowService");
-    // ServiceFactory.registerService("endService",
-    // "com.ly.train.flower.common.service.NothingService");
-
-    getServiceFlow("flower").buildFlow("flowService", "null");
-
-  }
 }
