+++ flower.sample/src/main/java/com/ly/train/flower/sample/web/service/UserServiceImpl.java
@@ -18,7 +18,7 @@ package com.ly.train.flower.sample.web.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.ly.train.flower.sample.web.dao.UserDao;
-import com.ly.train.flower.sample.web.mode.User;
+import com.ly.train.flower.sample.web.model.User;
 
 @Service("userService")
 public class UserServiceImpl {
