+++ flower.sample/src/main/java/com/ly/train/flower/sample/web/sync/SyncServlet.java
@@ -24,28 +24,29 @@ import javax.servlet.http.HttpServletResponse;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 import com.alibaba.fastjson.JSONObject;
-import com.ly.train.flower.sample.web.mode.User;
+import com.ly.train.flower.sample.web.model.User;
 import com.ly.train.flower.sample.web.service.UserServiceImpl;
 
 public class SyncServlet extends HttpServlet {
-  ApplicationContext context;
+	private static final long serialVersionUID = 1L;
+	ApplicationContext context;
 
-  @Override
-  public void init() {
-    context = new ClassPathXmlApplicationContext("spring-mybatis.xml");
-  }
+	@Override
+	public void init() {
+		context = new ClassPathXmlApplicationContext("spring-mybatis.xml");
+	}
 
-  @Override
-  protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-    resp.setContentType("text/html;charset=UTF-8");
-    int id = Integer.valueOf(req.getParameter("id").toString());
+	@Override
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		resp.setContentType("text/html;charset=UTF-8");
+		int id = Integer.valueOf(req.getParameter("id").toString());
 
-    UserServiceImpl userService = (UserServiceImpl) context.getBean("userService");
-    User user = userService.searchUser(id);
+		UserServiceImpl userService = (UserServiceImpl) context.getBean("userService");
+		User user = userService.searchUser(id);
 
-    String result = JSONObject.toJSONString(user);
-    PrintWriter out = resp.getWriter();
-    out.println(result);
-    out.flush();
-  }
+		String result = JSONObject.toJSONString(user);
+		PrintWriter out = resp.getWriter();
+		out.println(result);
+		out.flush();
+	}
 }
