+++ flower.sample/src/main/resources/application.yml
@@ -7,3 +7,8 @@ spring:
 
 server:
    port: 8081
+   
+flower:
+   name: flower-boot-demo
+   host: 127.0.0.1
+   port: 25002
